Initial Release
As of this release, the program can create the following EvoCreo
objects: Creo, Move, Trait, and Ability. These objects are loaded and
stored in a LoadingList, which has an individual vector for each object
type. LoadingList can also load each type of object from the game files
and return a constant pointer to that object on the list. The point of
the LoadingList is that each object only has to be loaded once, but
doesn't have to be loaded until it is requested.

The program must have the following files in the same directory when
run, unless different files are specified using LoadingList's "set"
functions: CreoData.xml, CreoStrings.xml, MoveData.xml, MoveStrings.xml,
and SceneStrings.xml. All of these are found in the app's files. NOTE:
There is a typo in CreoStrings.xml: "Shadiair" is spelled "Shadiar".
Running the program without correcting this typo will make it crash when
it tries to load "Shadiar", since there is no data for it, only strings,
and the very minimal exception handling I have put in place.

As of this release, only Creo objects can write pages.

To Do:
- Add Effect, Condition, and Boon structs
- Change ChanceEffect to use those structs
- Add function to write Move pages