Release 1.3
- Can now write Move pages, complete with lists of creo that can learn
the move by levelling up and by tome
- Move struct contains a vector called 'creoLearn', creo that can learn
that move, which is a list of LearnedSkill<Creo> sorted by creo ID
number.
- LearnedSkill has a new member, string tomeName.
- the LearnedSkill<Creo> in creoLearn use levelLearn and/or tomeName to
specify how a creo learns the move. levelLearned is 0 if it is not
learned by level, and tomeName is "" if it is not learned by tome. If a
creo can learn a move by both level and tome, that data will be stored
in one LearnedSkill.
- Creo.tomeMoves is now a vector of LearnedSkill<Move> using
LearnedSkill.tomeName.